+++ lib/main_db/providers/db_history_provider.dart
@@ -1,4 +1,4 @@
-import 'package:riverpod/riverpod.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../services/db_history_services/db_history_services.dart';